+++ packages/connectivity_status/example/test/widget_test.dart
@@ -33,6 +33,10 @@ class _FakeRepository implements ConnectivityRepository {
   final BehaviorSubject<ConnectivityStatus> _subject;
 
   @override
+  Stream<ConnectivityException> get exceptions =>
+      const Stream<ConnectivityException>.empty();
+
+  @override
   Future<ConnectivityStatus> getStatus() async => _subject.value;
 
   @override